Modify push script to create non-existent resource
fixes #128

When trying to update a resource on transifex, it catches the file not found response and tries to create it instead.

This required adding a function that will return the correct file type for the resource based on the repository. Note, this will fail for new repositories.

Also checking in package lock changes, although they appear to be just changing. `http` to `https`