remove obsolete conference 2020 resource
We "removed" the conference 2020 page (actually renamed it 2021) in www and in transifex. This doesn't break anything until a new language gets added.

When a new language is added there is no longer a resource in transifex for it to pull from, so there is no file for that language in the obsolete resource. Fortunately it fails the test phase when that happens. The solution is simply to remove the obsolete resource from l10n.